Add spec for vm evacuations for host recovery
When large scale hosts failed in the same time, there would be
not enough resources to recovery all the instances. So it needs
a priority to determine which one to evacuate firstly. Moreover,
the user even wants to recall the instances to the original host
after host recovery. So it needs to keep some necessary information
aboud evacuations for new features.

Change-Id: I7c87c8d4a2ab5898d1b0d046a1377fe1ed7c79ee
Blueprint: vm-evacuations-for-host-recovery